This requests pull of three commits:

* add TestPexInitialNoCutoff to catch regressions like anacrolix/torrent#433
* simplify `pexMsgFactory` to make better use of surrounding types
* optimise generation of the initial PEX

The latter deserves to be covered in greater detail.

The optimisation comes from the observation that every subsequent PEX message has to scan over the same bunch of events plus a small number of events on top of that. Included are benchmarks runnable with `go test -bench`.

Below are the benchmark results, and a plot showing the values of the last column along with the trend lines:

 * before:
```
BenchmarkPexInitial4-8     	  102764	     11319 ns/op
BenchmarkPexInitial50-8    	    2024	    500112 ns/op
BenchmarkPexInitial100-8   	     656	   1851580 ns/op
BenchmarkPexInitial200-8   	     159	   8022727 ns/op
BenchmarkPexInitial400-8   	      40	  30888163 ns/op
```

 * after:
```
BenchmarkPexInitial4-8     	  145088	      7539 ns/op
BenchmarkPexInitial50-8    	    6163	    187419 ns/op
BenchmarkPexInitial100-8   	    1693	    629353 ns/op
BenchmarkPexInitial200-8   	     465	   2236348 ns/op
BenchmarkPexInitial400-8   	     157	   7350410 ns/op
```

<img width="540" alt="PEXOptimisationPlot" src="https://user-images.githubusercontent.com/11537279/98701682-78ce6180-2371-11eb-9d53-957ddc6f9009.png">

[PEX optimisation.pdf](https://github.com/anacrolix/torrent/files/5518647/PEX.optimisation.pdf)