@yarikk 2020-11-10
@anacrolix here's the optimisation I had in mind – comments are welcome!