@yarikk 2020-11-12
@anacrolix Thanks for your feedback, much appreciated. I responded under each concern raised with explanations why I chose to go one way and not the other – I hope it clears things up. Also, good call adding maps to speed up the de-duplications – nice job!